Initial check-in
The code is reasonably solid, notwithstanding some inelegance which
needs refactoring in main().  On FreeBSD, it's been my dogfood for over
six weeks.  I have also built it and tried basic functions on Linux.

Some recent gems this found for my new tip addresses:

 - bc1qnullnymefa273hgss63kvtvr0q7377kjza0607
 - 35segwitgLKnDi2kn7unNdETrZzHD2c5xh

However, this started only as an afternoon project to meet my own needs.
For use by others, it requires serious improvement, especially in the
user interface and errors/warnings.  At present, "error handling" mostly
consists either of an abort(), or simple exit with a nonzero exit code.

The manpage (segvan.1) is partly written, but excluded from this
check-in because I consider the user interface to be as yet quite
unstable.