Remove secp256k1 directory to make way for subtree
The secp256k1 in the initial checkin was manually copied from Bitcoin
Core v0.15.1.  I will next replace it with a subtree of the new
sipa_201802_grind branch from nym-zone/secp256k1.

Apologies for the commit history noise.  I am doing this on master to
get the noise over with, and get it all done at once.